python_env/requirements.txt
# ATmega128 Educational Framework - Python Dependencies
# Assembly → C → Python → IoT Learning Progression

# Core communication
pyserial>=3.5

# Data analysis and visualization  
matplotlib>=3.5.0
numpy>=1.21.0
pandas>=1.3.0
seaborn>=0.11.0

# Optional: Enhanced plotting
plotly>=5.0.0

# Optional: Web interface (for future IoT demos)
flask>=2.0.0
flask-socketio>=5.0.0